css fixes

+++ app/api/auth_routes.py
@@ -55,8 +55,9 @@ def logout():
     Logs a user out
     """
     cart = Cart.query.filter(Cart.user_id == current_user.id).first()
-    db.session.delete(cart)
-    db.session.commit()
+    if cart:
+        db.session.delete(cart)
+        db.session.commit()
     logout_user()
     return {'message': 'User logged out'}
